Prevent CUDA fork deadlocks and fp16 overflows
Add multiple stability fixes and a config tweak:

- dataloader: set multiprocessing_context="spawn" when num_workers>0 and document why (prevents CUDA context inheritance deadlocks with pin_memory + DeepSpeed/CUDA).
- transformer: change TransformerOutput from @dataclass to NamedTuple so DeepSpeed ZeRO-3 can attach backward hooks (fixes an iteration-hang).
- mtp / trainer / perplexity: upcast logits to .float() before cross_entropy to avoid fp16 overflow in the per-token logsumexp across large vocab sizes. Comments explain the rationale.
- configs/125m.yaml: change grad_accum_steps 2->1 and micro_batch_size 4->8 (keeps effective batch size), and set data.num_workers: 1 to enable worker usage safely.

These changes address hangs and silent fp16 loss overflows while preserving training throughput by keeping the same effective batch size.

+++ ats/model/mtp.py
@@ -43,8 +43,12 @@ def compute_mtp_loss_from_logits(
             continue
         pred = logits[:, : seq_len - k, :].contiguous()
         target = labels[:, k:].contiguous()
+        # .float(): see ats/training/trainer.py's identical fix -- the
+        # per-token logsumexp reduction over vocab_size classes overflows
+        # fp16 purely from vocab_size (sums to >65504 regardless of
+        # prediction quality) unless logits are upcast first.
         loss_k = F.cross_entropy(
-            pred.reshape(-1, vocab_size),
+            pred.float().reshape(-1, vocab_size),
             target.reshape(-1),
             ignore_index=IGNORE_INDEX,
         )
